Added some casts to phong_frag_shader.glsl, needed for some drivers.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@8036 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ opengl/shaders/phong_frag_shader.glsl
@@ -118,7 +118,7 @@ void main()
 		sun_vis_factor = 0;
 		int pixel_index = int((gl_FragCoord.y * 1920.0 + gl_FragCoord.x));
 
-		float theta = ha(pixel_index) * (6.283185307179586 / 4294967296.0);
+		float theta = float(ha(uint(pixel_index)) * (6.283185307179586 / 4294967296.0));
 		mat2 R = mat2(cos(theta), sin(theta), -sin(theta), cos(theta));
 
 		for(int i=0; i<16; ++i)
